backend/requirements.txt
# Dependencies for the Smart Grocery + Recipe Planner
# Web Framework (Flask instead of Streamlit)
flask>=3.0.0
flask-cors>=4.0.0

# AI and API integrations
openai>=1.0.0
requests>=2.31.0

# Database support
firebase-admin>=6.2.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0  # Optional: For advanced database operations
pymongo>=4.0.0  # Optional: For MongoDB support

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Computer Vision (lightweight - only OpenCV)
opencv-python>=4.8.0
# Removed: dlib>=19.24.0 (too heavy, needs CMake)
# Removed: scipy>=1.10.0 (not needed for simple detection)
# Removed: imutils>=0.5.4 (not needed)
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0

# Additional Flask utilities
werkzeug>=3.0.0
jinja2>=3.1.0
